reorganize boundStore
