+++ src/store/utilsSlice.ts
@@ -1,25 +1,5 @@
-import { SliceWithMiddlewares } from '../typesStore';
-import { SettingsSlice } from './settingsSlice';
-import { BoundActions } from './boundStore';
+import { ActionsWithMiddlewares, UtilsState, UtilsActions, SettingsState, ControlsActions } from './types';
 import { CategoriesResponse, QuestionsResponse } from '../types';
-
-interface UtilsState {
-	isLoading: boolean;
-	isCountdown: boolean;
-	isModal: boolean;
-	timeLeft: number;
-	intervalId: number | null;
-}
-
-interface UtilsActions {
-	fetchWithRetry: (url: string, retries: number, backoff: number) => Promise<CategoriesResponse | QuestionsResponse | undefined>;
-	toggleCountdown: () => void;
-	setIsModal: (isModal: boolean) => void;
-	runIntervalId: (timer: number) => void;
-	clearIntervalId: () => void;
-}
-
-export interface UtilsSlice extends UtilsState, UtilsActions {}
 
 export const initialUtilsState: UtilsState = {
 	isLoading: false,
@@ -29,14 +9,12 @@ export const initialUtilsState: UtilsState = {
 	intervalId: null,
 }
 
-export const createUtilsSlice: SliceWithMiddlewares<
-UtilsSlice & SettingsSlice & BoundActions,
-UtilsSlice
+export const createUtilsActions: ActionsWithMiddlewares<
+UtilsState & Pick<UtilsActions, 'toggleCountdown' | 'clearIntervalId'> & SettingsState & Pick<ControlsActions, 'handleNextButton'>,
+UtilsActions
 > = (set, get) => ({
-	...initialUtilsState,
-
 	fetchWithRetry: async (url, retries, backoff) => {
-		set({ isLoading: true }, false, 'utils/setIsLoadingTrue');
+		set({ isLoading: true }, undefined, 'utils/setIsLoadingTrue');
 		for (let i = 0; i < retries; i++) {
 			try {
 				const response = await fetch(url);
@@ -52,25 +30,29 @@ UtilsSlice
 					throw error;
 				}
 			} finally {
-				set({ isLoading: false }, false, 'utils/setIsLoadingFalse');
+				set({ isLoading: false }, undefined, 'utils/setIsLoadingFalse');
 			}
 		}
 	},
 
 	toggleCountdown: () => {
-		set((state) => ({ isCountdown: !state.isCountdown }), false, 'utils/toggleCountdown');
+		set((state) => ({ isCountdown: !state.isCountdown }),
+			undefined,
+			'utils/toggleCountdown');
 	},
 
 	setIsModal: (isModal) => {
-		set({ isModal }, false, 'utils/setIsModal');
+		set({ isModal },
+			undefined,
+			'utils/setIsModal');
 	},
 
 	runIntervalId: (timer) => {
 		if (get().intervalId === null) {
-			set({ timeLeft: timer }, false, 'quiz/setInitTimeLeft');
+			set({ timeLeft: timer }, undefined, 'quiz/setInitTimeLeft');
 
 			get().intervalId = setInterval(() => {
-				set((state) => { state.timeLeft -= 1; }, false, 'quiz/decTimeLeft');
+				set((state) => { state.timeLeft -= 1; }, undefined, 'quiz/decTimeLeft');
 				if (get().timeLeft === 0) {
 					if (get().settings.timer && !get().isCountdown) {
 						get().handleNextButton();
@@ -92,6 +74,7 @@ UtilsSlice
 				state.intervalId = null;
 			}
 		},
-		false, 'quiz/clearIntervalId');
+		undefined,
+		'quiz/clearIntervalId');
 	}
 });
