--- src/typesStore.ts
@@ -1,8 +1,0 @@
-import { StateCreator } from 'zustand';
-
-export type SliceWithMiddlewares<T, S = T> = StateCreator<
-    T,
-    [["zustand/devtools", never], ["zustand/immer", never]],
-    [],
-    S extends T ? T : S
->;
